Adding protection

+++ app/Http/Controllers/NoteController.php
@@ -3,45 +3,51 @@
 namespace App\Http\Controllers;
 
 use App\Note;
-use App\NotesImage;
-use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class NoteController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth');
+    }
+
     public function index(){
-      $notes = Note::all();
-      return view('note.index', compact('notes'));
+        $notes = Note::all()->where('user_id' , Auth::id());
+        return view('note.index', compact('notes'));
     }
 
     public function create(){
-      return view('note.create');
+        return view('note.create');
     }
 
     public function store(){
-      $note = Note::create(Request()->validate([
-        'title' => 'required|min:5',
-        'description' => 'required|min:10'
-      ]));
-      $notes = Note::all();
-      $messege = "Dodano nową notatkę.";
-      return redirect('/');
+        $validated = Request()->validate([
+            'title' => 'required|min:5',
+            'description' => 'required|min:10'
+        ]);
+        $validated['user_id'] = Auth::id();
+        Note::create($validated);
+        return redirect('/');
     }
 
     public function edit(Note $note){
-      return view('note.edit', compact('note'));
+        if($note->isNotAuthor()) abort(403);
+        return view('note.edit', compact('note'));
     }
 
     public function update(Note $note){
-      $validatedNote = Request()->validate([
-        'title' => 'required|min:5',
-        'description' => 'required|min:10'
-      ]);
-      $note->update($validatedNote);
-      return redirect('/');
+        if($note->isNotAuthor()) abort(403);
+        $validatedNote = Request()->validate([
+            'title' => 'required|min:5',
+            'description' => 'required|min:10'
+        ]);
+        $note->update($validatedNote);
+        return redirect('/');
     }
 
     public function destroy(Note $note){
-        $note->images()->delete();
+        if($note->isNotAuthor()) abort(403);
         $note->delete();
         return redirect('/');
     }
